chore: Refactor button proof of concept
