+++ modules/button/react/lib/iconButton/iconButtonTypes.ts
@@ -1,0 +1,14 @@
+export enum IconButtonVariant {
+  Square = 'square',
+  SquareFilled = 'squareFilled',
+  Plain = 'plain',
+  Circle = 'circle',
+  CircleFilled = 'circleFilled',
+  Inverse = 'inverse',
+  InverseFilled = 'inverseFilled',
+}
+
+export enum IconButtonSize {
+  Small = 'small',
+  Medium = 'medium',
+}
