+++ modules/button/react/lib/utils.ts
@@ -6,7 +6,7 @@ import {
   AllButtonVariants,
   TextButtonVariant,
   ButtonVariant,
-  IconButtonVariant,
+  // IconButtonVariant,
 } from './types';
 import * as ButtonStyles from './ButtonStyles';
 import {ButtonColors} from './ButtonColors';
@@ -61,20 +61,20 @@ export function getButtonStyle(
       return types[TextButtonVariant.AllCaps];
     case TextButtonVariant.InverseAllCaps:
       return types[TextButtonVariant.InverseAllCaps];
-    case IconButtonVariant.Square:
-      return types[IconButtonVariant.Square];
-    case IconButtonVariant.SquareFilled:
-      return types[IconButtonVariant.SquareFilled];
-    case IconButtonVariant.Plain:
-      return types[IconButtonVariant.Plain];
-    case IconButtonVariant.Circle:
-      return types[IconButtonVariant.Circle];
-    case IconButtonVariant.CircleFilled:
-      return types[IconButtonVariant.CircleFilled];
-    case IconButtonVariant.Inverse:
-      return types[IconButtonVariant.Inverse];
-    case IconButtonVariant.InverseFilled:
-      return types[IconButtonVariant.InverseFilled];
+    // case IconButtonVariant.Square:
+    //   return types[IconButtonVariant.Square];
+    // case IconButtonVariant.SquareFilled:
+    //   return types[IconButtonVariant.SquareFilled];
+    // case IconButtonVariant.Plain:
+    //   return types[IconButtonVariant.Plain];
+    // case IconButtonVariant.Circle:
+    //   return types[IconButtonVariant.Circle];
+    // case IconButtonVariant.CircleFilled:
+    //   return types[IconButtonVariant.CircleFilled];
+    // case IconButtonVariant.Inverse:
+    //   return types[IconButtonVariant.Inverse];
+    // case IconButtonVariant.InverseFilled:
+    //   return types[IconButtonVariant.InverseFilled];
   }
 }
 
@@ -111,11 +111,11 @@ export function getButtonFocusRing(variant: AllButtonVariants): CSSObject {
       return focusRing(2, 0);
     case ButtonVariant.OutlineInverse:
       return focusRing(2, 2, true, false, buttonColors.focusRingInner, buttonColors.focusRingOuter);
-    case IconButtonVariant.Plain:
-      return focusRing(2);
-    case IconButtonVariant.Inverse:
-    case IconButtonVariant.InverseFilled:
-      return focusRing(2, 2, true, false, buttonColors.focusRingInner, buttonColors.focusRingOuter);
+    // case IconButtonVariant.Plain:
+    //   return focusRing(2);
+    // case IconButtonVariant.Inverse:
+    // case IconButtonVariant.InverseFilled:
+    //   return focusRing(2, 2, true, false, buttonColors.focusRingInner, buttonColors.focusRingOuter);
     case TextButtonVariant.Inverse:
     case TextButtonVariant.InverseAllCaps:
       return focusRing(2, 0, true, false, buttonColors.focusRingInner, buttonColors.focusRingOuter);
@@ -243,114 +243,114 @@ export function getButtonStateStyle(variant: AllButtonVariants): CSSObject {
   };
 }
 
-export function getIconButtonStateStyle(variant: AllButtonVariants): CSSObject {
-  const buttonColors = ButtonColors[variant];
+// export function getIconButtonStateStyle(variant: AllButtonVariants): CSSObject {
+//   const buttonColors = ButtonColors[variant];
 
-  if (buttonColors == null) {
-    return {};
-  }
+//   if (buttonColors == null) {
+//     return {};
+//   }
 
-  const baseStyles = {
-    borderColor: buttonColors.border,
-    backgroundColor: buttonColors.background,
-    ...(buttonColors.labelIcon && {
-      'span .wd-icon-fill, span .wd-icon-accent': {
-        transition: 'fill 120ms ease-in',
-        fill: buttonColors.labelIcon,
-      },
-    }),
-    ...(buttonColors.labelData && {
-      ['.' + ButtonStyles.labelDataBaseStyles.classname]: {
-        color: buttonColors.labelData,
-      },
-    }),
-  };
+//   const baseStyles = {
+//     borderColor: buttonColors.border,
+//     backgroundColor: buttonColors.background,
+//     ...(buttonColors.labelIcon && {
+//       'span .wd-icon-fill, span .wd-icon-accent': {
+//         transition: 'fill 120ms ease-in',
+//         fill: buttonColors.labelIcon,
+//       },
+//     }),
+//     ...(buttonColors.labelData && {
+//       ['.' + ButtonStyles.labelDataBaseStyles.classname]: {
+//         color: buttonColors.labelData,
+//       },
+//     }),
+//   };
 
-  const hoverStyles = {
-    ':hover': {
-      backgroundColor: buttonColors.hoverBackground,
-      ...(buttonColors.labelIconHover && {
-        'span .wd-icon-fill, span .wd-icon-accent': {
-          fill: buttonColors.labelIconHover,
-        },
-      }),
-    },
-  };
+//   const hoverStyles = {
+//     ':hover': {
+//       backgroundColor: buttonColors.hoverBackground,
+//       ...(buttonColors.labelIconHover && {
+//         'span .wd-icon-fill, span .wd-icon-accent': {
+//           fill: buttonColors.labelIconHover,
+//         },
+//       }),
+//     },
+//   };
 
-  const activeStyles = {
-    ':active, :focus:active, :hover:active': {
-      backgroundColor: buttonColors.activeBackground,
-      ...(buttonColors.labelIconHover && {
-        'span .wd-icon-fill, span .wd-icon-accent': {
-          fill: buttonColors.labelIconActive,
-        },
-      }),
-    },
-  };
+//   const activeStyles = {
+//     ':active, :focus:active, :hover:active': {
+//       backgroundColor: buttonColors.activeBackground,
+//       ...(buttonColors.labelIconHover && {
+//         'span .wd-icon-fill, span .wd-icon-accent': {
+//           fill: buttonColors.labelIconActive,
+//         },
+//       }),
+//     },
+//   };
 
-  return {
-    ...baseStyles,
-    ':focus': {
-      backgroundColor: buttonColors.focusBackground,
-      ...(buttonColors.labelDataFocus && {
-        ['.' + ButtonStyles.labelDataBaseStyles.classname]: {
-          color: buttonColors.labelDataFocus,
-        },
-      }),
-      ...(buttonColors.labelIconFocus && {
-        'span .wd-icon-fill, span .wd-icon-accent': {
-          fill: buttonColors.labelIconFocus,
-        },
-      }),
-    },
-    ...hoverStyles,
-    ...activeStyles,
-    ':disabled, :active:disabled, :focus:disabled, :hover:disabled': {
-      pointerEvents: 'none',
-      backgroundColor: buttonColors.disabledBackground,
-      borderColor: buttonColors.disabledBorder,
-      color: buttonColors.disabledText,
-      ...(buttonColors.labelIconDisabled && {
-        'span .wd-icon-fill, span .wd-icon-accent': {
-          fill: buttonColors.labelIconDisabled,
-        },
-      }),
-      ...(buttonColors.labelDataDisabled && {
-        ['.' + ButtonStyles.labelDataBaseStyles.classname]: {
-          color: buttonColors.labelDataDisabled,
-        },
-      }),
-    },
-    '&:not([disabled])': {
-      '&:focus': {
-        background: buttonColors.focusBackground,
-        borderColor: buttonColors.focusBorder,
-        ...getButtonFocusRing(variant),
-      },
-      '&:active': {
-        borderColor: buttonColors.activeBorder,
-        ...getButtonFocusRing(variant),
-      },
+//   return {
+//     ...baseStyles,
+//     ':focus': {
+//       backgroundColor: buttonColors.focusBackground,
+//       ...(buttonColors.labelDataFocus && {
+//         ['.' + ButtonStyles.labelDataBaseStyles.classname]: {
+//           color: buttonColors.labelDataFocus,
+//         },
+//       }),
+//       ...(buttonColors.labelIconFocus && {
+//         'span .wd-icon-fill, span .wd-icon-accent': {
+//           fill: buttonColors.labelIconFocus,
+//         },
+//       }),
+//     },
+//     ...hoverStyles,
+//     ...activeStyles,
+//     ':disabled, :active:disabled, :focus:disabled, :hover:disabled': {
+//       pointerEvents: 'none',
+//       backgroundColor: buttonColors.disabledBackground,
+//       borderColor: buttonColors.disabledBorder,
+//       color: buttonColors.disabledText,
+//       ...(buttonColors.labelIconDisabled && {
+//         'span .wd-icon-fill, span .wd-icon-accent': {
+//           fill: buttonColors.labelIconDisabled,
+//         },
+//       }),
+//       ...(buttonColors.labelDataDisabled && {
+//         ['.' + ButtonStyles.labelDataBaseStyles.classname]: {
+//           color: buttonColors.labelDataDisabled,
+//         },
+//       }),
+//     },
+//     '&:not([disabled])': {
+//       '&:focus': {
+//         background: buttonColors.focusBackground,
+//         borderColor: buttonColors.focusBorder,
+//         ...getButtonFocusRing(variant),
+//       },
+//       '&:active': {
+//         borderColor: buttonColors.activeBorder,
+//         ...getButtonFocusRing(variant),
+//       },
 
-      '&:hover:focus': {
-        backgroundColor: buttonColors.focusHover,
-        'span .wd-icon-fill, span .wd-icon-accent': {
-          fill: buttonColors.labelIconFocusHover,
-        },
-      },
-      '&:focus:active': {
-        backgroundColor: buttonColors.activeBackground,
-      },
-    },
-    ...mouseFocusBehavior({
-      '&:focus': {
-        ...baseStyles,
-        outline: 'none',
-        boxShadow: 'none',
-        animation: 'none',
-        ...hoverStyles,
-        ...activeStyles,
-      },
-    }),
-  };
-}
+//       '&:hover:focus': {
+//         backgroundColor: buttonColors.focusHover,
+//         'span .wd-icon-fill, span .wd-icon-accent': {
+//           fill: buttonColors.labelIconFocusHover,
+//         },
+//       },
+//       '&:focus:active': {
+//         backgroundColor: buttonColors.activeBackground,
+//       },
+//     },
+//     ...mouseFocusBehavior({
+//       '&:focus': {
+//         ...baseStyles,
+//         outline: 'none',
+//         boxShadow: 'none',
+//         animation: 'none',
+//         ...hoverStyles,
+//         ...activeStyles,
+//       },
+//     }),
+//   };
+// }
